requirements.txt
# Core dependencies
streamlit>=1.31.0
streamlit_chat==0.1.1
streamlit-extras>=0.3.6
streamlit-ace>=0.1.1
streamlit-elements>=0.1.0

# Data processing
polars>=0.20.2
pandas>=2.1.4
numpy>=1.26.3
pyarrow>=14.0.2
requests>=2.31.0
Pillow>=10.1.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.1
plotly>=5.18.0
altair>=5.2.0

# Machine learning
scikit-learn>=1.4.0

# Database dependencies
sqlalchemy>=2.0.27
sqlmodel>=0.0.14
pymysql>=1.1.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0

# LangChain dependencies
langchain>=0.1.4
langchain-openai>=0.0.5
langchain-community>=0.0.19
langchain-huggingface>=0.0.8

# PDF processing
pypdf>=3.17.1
unstructured>=0.11.2
tabula-py>=2.9.0

# Excel processing
openpyxl>=3.1.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.1
httpx>=0.26.0
rich>=13.7.0
tqdm>=4.66.1
